// planner.ui/src/types/message-key.ts
export enum MessageKey {
  ServerError = 0,
  Title = 1,
  Description = 2,
  DueDate = 3,
  Save = 4,
  ErrorTitle = 5,
  GoalFormTitle = 6,
  GoalListTitle = 7,
  Delete = 8,
  Complete = 9,
  Add = 10,
  Edit = 11,
  Cancel = 12,
  SaveAndClose = 13,
  Goal_Add_Button = 14,
  Goal_Edit_Button = 15,
  Goal_Delete_Confirm = 16,
  Goal_Completed_Message = 17,
  Goals_Empty = 18,
  YearlyGoals_Title = 19,
  YearlyGoal_Add = 20,
  YearlyGoal_Edit = 21,
  YearlyGoal_Delete_Confirm = 22,
  Field_StartDate = 23,
  Field_DueDate = 24,
  Field_Priority = 25,
  Field_Tags = 26,
  Unsaved_Changes = 27,
  Max_YearlyGoals_Message = 28,
  Toast_Goal_Created = 29,
  Toast_Goal_Updated = 30,
  Toast_YearlyGoal_Added = 31,
  Toast_YearlyGoal_Updated = 32,
  Toast_YearlyGoal_Deleted = 33,
  Toast_Goal_Deleted = 34,
  Back = 35,
  Goal_NotFound = 36,
  YearlyGoal_NotFound = 37,
  Invalid_Input = 38,
  Operation_Failed = 39,
  DailyGoal_NotFound = 40,
  MonthlyGoal_NotFound = 41,
  WeeklyGoal_NotFound = 42,
  Loading = 43,
  NoGoals = 44,
  Completed = 45,
  YearlyGoals = 46,
  Target = 47,
  Plan = 48,
  Close = 49,
  GoalTreeTitle = 50,
  Yearly = 51,
  Monthly = 52,
  Weekly = 53,
  Daily = 54,
  NoYearlyGoals = 55,
  DeleteTreeItemConfirm = 56,
  DeleteTreeItemWithChildrenConfirm = 57,
  Language = 58,
  Year = 59,
  Month = 60,
  Day = 61,
}